move into namespace, implement Iplaceable into building placement
