+++ LD47/Assets/Scripts/Behaviour/PlayerBuilding.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7af00bd10c4f3054c8523d1e93d6075a
+guid: 8f89054ebe6468e4abae544625ebce3b
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
