+++ LD47/Assets/Scripts/Behaviour/PlayerBuilding/IBuildable.cs
@@ -1,0 +1,7 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public interface IBuildable {
+
+}
