+++ LD47/Assets/Scripts/Behaviour/PlayerBuilding/IBuildable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 670adda7792ef3c43a1f66c5fb8f6f6d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
